requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
markdownify>=0.11.6
lxml>=5.2.0
PyYAML>=6.0.1
